+++ App/DISClib/DataStructures/linkedlistiterator.py
