+++ App/DISClib/DataStructures/mapentry.py
