+++ main.py
@@ -1,0 +1,3 @@
+from App import view
+
+view.run()
